Created graphUtils and main packages
- Created utility for reading graph from text file
	- See "exampleGraph.txt" for format
- Created toString method to Graph
- Implemented initialisation for Vertex fields
- Changed adjacency list weights to float
- Updated timelog.md

+++ GraphCompression/src/graphComponents/AdjNode.java
@@ -4,9 +4,9 @@ package graphComponents;
 /*Represents an adjacent Vertex in the adjacency list of a Vertex*/
 public class AdjNode {
 	private int index;	//Index of the Vertex represented by this node in the Graph's Vertex list
-	private int weight;	//Represents the weight of the edge to get from some Vertex to this adjacent Vertex
+	private float weight;	//Represents the weight of the edge to get from some Vertex to this adjacent Vertex
 	
-	public AdjNode(int index, int weight) {
+	public AdjNode(int index, float weight) {
 		this.index  = index;
 		this.weight = weight;
 	}
@@ -19,7 +19,7 @@ public class AdjNode {
 	
 	
 	/*Returns weight of the edge to reach this Vertex from some adjacent Vertex*/
-	public int getWeight() {
+	public float getWeight() {
 		return this.weight;
 	}
 }
